# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "auto-reel-ng API",
    "version": "1"
  },
  "paths": {
...
    "/api/v1/events": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "Get Events",
        "description": "``GET /api/v1/events`` (task 2.2): every event, freshly scanned.\n\nOne unreadable event costs one error row, never the list (``list_events``\nisolates per event). What remains are the whole-list failures, reported\ndistinctly in the shared problem shape: an unreachable job store as the 503\n``/healthz`` returns, a walk of the project root that fails as the scan-failure\n502. Neither is retried and neither degrades into a partial list.",
        "operationId": "get_events_api_v1_events_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "oneOf": [
                      {
                        "$ref": "#/components/schemas/EventSummaryOut"
                      },
                      {
                        "$ref": "#/components/schemas/EventErrorOut"
                      }
                    ],
                    "discriminator": {
                      "propertyName": "kind",
                      "mapping": {
                        "event": "#/components/schemas/EventSummaryOut",
                        "error": "#/components/schemas/EventErrorOut"
                      }
                    }
                  },
                  "type": "array",
                  "title": "Response Get Events Api V1 Events Get"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/events/{event_id}/analysis": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "Get Analysis",
        "description": "``GET /api/v1/events/{event_id}/analysis`` (task 2.4): cached segments only.\n\nRegistered *before* the ``{event_id:path}`` detail route below: both patterns\nare greedy over ``/``, and Starlette matches routes in registration order, so\nthe more specific ``/analysis`` suffix must be tried first or the detail route\nwould swallow it (``event_id`` ending in literal ``/analysis``).",
        "operationId": "get_analysis_api_v1_events__event_id__analysis_get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/events/{event_id}/reel": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "Get Reel",
        "description": "``GET /api/v1/events/{event_id}/reel``: the editorial document as authored.\n\nThe exact inverse of the PUT below — same model, same serializer — so a client\nmay submit what it read without hand-building a write body. Registered *before*\nthe greedy ``{event_id:path}`` detail route for the reason the ``/analysis``\nroute documents. The response carries an ``ETag`` for the optional ``If-Match``\nprecondition on the write; there is deliberately no conditional GET (D-R4).",
        "operationId": "get_reel_api_v1_events__event_id__reel_get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditorialDocumentBody-Output"
                }
              }
            },
            "headers": {
              "ETag": {
                "description": "Strong entity-tag of the editorial state, for If-Match on the write",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Not Found"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "events"
        ],
        "summary": "Put Reel",
        "description": "``PUT /api/v1/events/{event_id}/reel``: apply a desired editorial state (D-E2).\n\nA coarse whole-document write: ``payload`` is the complete desired editorial\nstate, merged onto the event's existing structure (never replacing the file\noutright, D-E1) and validated fail-loud before anything is persisted. The\nroute carries no editorial logic of its own (D-E6) — it resolves the event,\ndelegates to the engine operation, and echoes the persisted document plus the\nevent's resulting staleness verdict (computed free, D-E5) so the client needs\nno follow-up GET. The write never enqueues and never touches the render\nmanifest.\n\nA successful write carries an ``ETag`` for the state it just persisted,\ncomputed from the returned document by the same helper ``get_reel`` uses, so a\nclient may chain conditional writes with no intervening read. A ``412`` carries\nnone: a client that lost the race must re-read before it overwrites.\n\nFailures answer by cause and write nothing: 400 for an invalid submitted state\n(with ``failure: unusable_metadata`` when the engine refuses a state that\nleaves the event without a real date or title), 404, 412, and 502 for the disk:\nan existing document that cannot be read (with its ``failure`` kind, as the\nreads report it) or a save the filesystem refuses (naming the OS error).",
        "operationId": "put_reel_api_v1_events__event_id__reel_put",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditorialDocumentBody-Input"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditorialWriteResult"
                }
              }
            },
            "headers": {
              "ETag": {
                "description": "Strong entity-tag of the editorial state, for If-Match on the write",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Not Found"
          },
          "412": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Precondition Failed"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/events/{event_id}": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "Get Event",
        "description": "``GET /api/v1/events/{event_id}`` (task 2.3): current detail from disk.\n\nAn event the list would show as an error row is the scan-failure 502 here,\ncarrying the same ``failure`` kind and ``detail`` as that row, never a 500.",
        "operationId": "get_event_api_v1_events__event_id__get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventDetailOut"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Not Found"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EventErrorOut": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "error",
            "title": "Kind"
          },
          "event_id": {
            "type": "string",
            "title": "Event Id"
          },
          "failure": {
            "$ref": "#/components/schemas/EventFailure"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "event_id",
          "failure",
          "detail"
        ],
        "title": "EventErrorOut",
        "description": "An event the list could not read, in place of its summary (per-event isolation).\n\nCarries no clip counts, staleness or title: those are exactly what could not be\nread, and absence is reported, never faked as zero (Principle I)."
      },
...
      "EventSummaryOut": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "event",
            "title": "Kind"
          },
          "event_id": {
            "type": "string",
            "title": "Event Id"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "clip_count": {
            "type": "integer",
            "title": "Clip Count"
          },
          "new_count": {
            "type": "integer",
            "title": "New Count"
          },
          "missing_count": {
            "type": "integer",
            "title": "Missing Count"
          },
          "latest_job": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobSummaryOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "staleness": {
            "$ref": "#/components/schemas/StalenessOut"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "event_id",
          "clip_count",
          "new_count",
          "missing_count",
          "staleness"
        ],
        "title": "EventSummaryOut",
        "description": "One event as listed by ``GET /api/v1/events``."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  }
}